filtro coordenador
