+++ src/main/java/org/ged/usuario/service/UsuarioService.java
@@ -1,0 +1,15 @@
+package org.ged.usuario.service;
+
+import jakarta.ws.rs.core.Response;
+import org.ged.usuario.api.v1.request.UsuarioRequest;
+import org.ged.usuario.entity.UsuarioEntity;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+public interface UsuarioService {
+
+    UsuarioEntity save(UsuarioRequest request);
+
+    List<UsuarioEntity> findByFilter(UsuarioRequest filter);
+}
